Hackaton08/AULLOA/requirements.txt
backpack==0.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
cleo==0.6.8
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
Faker==0.8.18
Flask==2.3.1
flask-orator==0.2.0
flask-swagger-ui==4.11.1
httpie==3.2.1
idna==3.4
inflection==0.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
psycopg2==2.9.6
pyaml==16.12.2
Pygments==2.15.1
pylev==1.4.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.29.0
requests-toolbelt==0.10.1
rich==13.3.5
simplejson==3.19.1
six==1.16.0
text-unidecode==1.2
tzlocal==1.5.1
urllib3==1.26.15
Werkzeug==2.3.1
wrapt==1.15.0
dnspython==2.3.0
python-dotenv==1.0.0